Reference

Algorithm

IterativeLQR.backward_pass! — Method

Perform the backward pass of the algorithm. Mainly, produce the the terms d and K of the policy update as well as the expected improvement Δv.

IterativeLQR.cost_regularization! — Method

Regularize the hessians of the running cost l and final cost Φ. (multithreaded)

IterativeLQR.differentiation! — Method

Produce partial derivatives of the system's dynamics f and costs l and Φ that are required for the backward pass of the algorithm. (multithreaded)

IterativeLQR.differentiation! — Method

Produce partial derivatives of the system's dynamics f and costs l and Φ that are required for the backward pass of the algorithm. Also transform them to the surface of the state manifold using coordinate jacobians. (multithreaded)

IterativeLQR.forward_pass! — Method

Perform the forward pass of the algorithm, i.e. produce a candidate trajectory x, u.

IterativeLQR.iLQR! — Method

Optimize the trajectory of a system with in-place dynamics! for the running_cost and final_cost on a horizon of length N. The number of the system's states nx and inputs nu as well as the horizon's length N must be compatible with the workset which is constructed using these quantities.

Functions dynamics_diff!, running_cost_diff!, and final_cost_diff! are used to calculate partial derivatives. They can either calculate each partial derivatives with respect to x and u separately:

dynamics_diff!(fx, fu, x, u, k),

running_cost_diff!(lx, lu, lxx, lxu, luu, x, u, k),

final_cost_diff!(Φx, Φxx, x, k)

or in a stacked form:

dynamics_diff!(∇f, x, u, k),

running_cost_diff!(∇l, ∇2l, x, u, k),

final_cost_diff!(Φx, Φxx, x, k).

If the stacked form is used, set the keyword argument stacked_derivatives to True.

If the state of the system lies on a manifold, the keyword arguments state_difference and coordinate_jacobian must be supplied. Also ndx!=nx must have been used in the construction of workset.

To re-use the nominal control policy from a previous solve (for example in MPC applications) set rollout=:partial. The default rollout=:full uses the nominal inputs in feedforward manner to produce the states of the nominal trajectory.

During the optimization, printout into the console can be switched of using by setting verbose=false. Instead of printing (or in addition to) information about each iteration can be stored in a dataframe::DataFrames.DataFrame by setting logging=true. If set true the dataframe is the only output of the function.

Quantities in the workset can be plotted live using the plotting_callback keyword argument which expects a function with the signature plotting_callback(workset). It is called once after the intial rollout and then after every succesful iteration.

IterativeLQR.iteration_dataframe — Method

Create a data frame for logging information about the iterations of the algorithm.

IterativeLQR.log_iteration! — Method

Log information into the data frame about the iterations of the algorithm.

IterativeLQR.print_iteration! — Method

Print information about the current iteration.

IterativeLQR.trajectory_rollout! — Method

Roll-out the states x̄ₖ of the nominal trajectory from the initial state x̃₁ and nominal inputs ūₖ. Also evaluate the running costs lₖ in multiple threads and the final cost Φ.

Workset

IterativeLQR.BackwardPassWorkset — Type

Struct containing quantities used exclusively in the backward pass of the algorithm.

IterativeLQR.CoordinateJacobians — Type

Struct containing coordinate jacobians Eₖ.

IterativeLQR.CostDerivatives — Type

Struct containing partial derivatives of the running cost l and the final cost Φ.

IterativeLQR.DynamicsDerivatives — Type

Struct containing partial derivatives of the system's dyanamics f.

IterativeLQR.PolicyUpdate — Type

Struct containing terms dₖ and Kₖ of the policy update δuₖ(δxₖ) = dₖ + Kₖδxₖ.

IterativeLQR.Trajectory — Type

Struct containing a trajectory xₖ, uₖ of the controlled system (including costs).

IterativeLQR.Workset — Type

Struct containing the nominal trajectory (current the most optimal) as well as all internal quantities of the algorithm.

IterativeLQR.Workset — Method

Create a Workset for the trajectory optimization problem.

Arguments

nx: number of states

nu: number of inputs

N: length of the optimizations horizon

ndx: number of independent states (defaults to nx)

Returns

IterativeLQR.active_trajectory — Method

Access the active trajectory

IterativeLQR.nominal_trajectory — Method

Access the nominal trajectory

IterativeLQR.set_initial_state! — Method

Set the initial state x̃₁.

IterativeLQR.set_nominal_inputs! — Function

Set the nominal inputs ūₖ.

IterativeLQR.swap_trajectories! — Method

Swap the nominal and active trajectory (without copying any data).

Regularization

IterativeLQR.eigenvalue_regularization! — Method

Regularizes H using the eigenvalue approach, so that all of its eigen values are >=δ.

IterativeLQR.gmw_regularization! — Method

Regularized H using the Gill-Murray-Wright modified Cholesky factorization.

IterativeLQR.regularize! — Method

Regularizes the symmetric matrix H so that it is positive definite. The argument approach determines if an eigenvalue approach is taken eigenvalue_regularization! or the GMW modified Cholesky decomposition algorithm is used gmw_regularization!.